# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mission Control API for openclaw agent Ben
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: agent
    description: Agent status and control
  - name: tasks
    description: Task management
  - name: logs
    description: Activity logs
  - name: memory
    description: Agent memory and context
  - name: projects
    description: Project tracking
  - name: events
    description: Calendar events and agenda
  - name: inbox
    description: Quick capture inbox
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /agent:
    get:
      operationId: getAgentStatus
      tags: [agent]
      summary: Get agent status
      description: Returns Ben's current status and metadata
      responses:
        "200":
          description: Agent status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentStatus"

  /agent/control:
    post:
      operationId: controlAgent
      tags: [agent]
      summary: Control the agent
      description: Start, stop, pause, or resume the agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentControlRequest"
      responses:
        "200":
          description: Control action result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentStatus"

  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: List all tasks
      description: Returns all tasks in the queue
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      description: Adds a new task to Ben's queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      description: Update a task's status or details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      description: Remove a task from the queue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Task deleted

  /logs:
    get:
      operationId: getLogs
      tags: [logs]
      summary: Get activity logs
      description: Returns recent activity logs for Ben
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: level
          in: query
          schema:
            type: string
            enum: [info, warning, error, success, all]
            default: all
      responses:
        "200":
          description: Activity logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogEntry"
    post:
      operationId: createLog
      tags: [logs]
      summary: Create a log entry
      description: Add a new log entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLogRequest"
      responses:
        "201":
          description: Log entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogEntry"

  /memory:
    get:
      operationId: getMemory
      tags: [memory]
      summary: Get agent memory
      description: Returns Ben's current memory and context items
      responses:
        "200":
          description: Memory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryItem"
    post:
      operationId: createMemory
      tags: [memory]
      summary: Create memory item
      description: Store a new memory or context item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemoryRequest"
      responses:
        "201":
          description: Memory item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryItem"

  /memory/{id}:
    delete:
      operationId: deleteMemory
      tags: [memory]
      summary: Delete memory item
      description: Remove a memory item from Ben's context
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Memory item deleted

  /events:
    get:
      operationId: getEvents
      tags: [events]
      summary: List calendar events
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Event list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEvent"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"

  /events/{id}:
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventRequest"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Event deleted

  /inbox:
    get:
      operationId: getInbox
      tags: [inbox]
      summary: List inbox items
      responses:
        "200":
          description: Inbox items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InboxItem"
    post:
      operationId: createInboxItem
      tags: [inbox]
      summary: Capture an inbox item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInboxItemRequest"
      responses:
        "201":
          description: Item captured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxItem"

  /inbox/{id}:
    patch:
      operationId: updateInboxItem
      tags: [inbox]
      summary: Update inbox item (e.g. mark processed)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInboxItemRequest"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxItem"
    delete:
      operationId: deleteInboxItem
      tags: [inbox]
      summary: Delete inbox item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Item deleted

  /projects:
    get:
      operationId: getProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}:
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Project deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    AgentStatus:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        state:
          type: string
          enum: [active, idle, paused, stopped, error]
        currentTask:
          type: string
          nullable: true
        uptime:
          type: integer
          description: Seconds since last start
        taskCount:
          type: integer
        completedCount:
          type: integer
        errorCount:
          type: integer
        lastActive:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - state
        - taskCount
        - completedCount
        - errorCount
        - updatedAt

    AgentControlRequest:
      type: object
      properties:
        action:
          type: string
          enum: [start, stop, pause, resume]
      required:
        - action

    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, in_progress, completed, failed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, critical]
        dueDate:
          type: string
          nullable: true
        projectId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - title
        - status
        - priority
        - createdAt
        - updatedAt

    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        priority:
          type: string
          enum: [low, medium, high, critical]
          default: medium
        dueDate:
          type: string
          nullable: true
        projectId:
          type: integer
          nullable: true
      required:
        - title

    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, in_progress, completed, failed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, critical]
        dueDate:
          type: string
          nullable: true
        projectId:
          type: integer
          nullable: true

    LogEntry:
      type: object
      properties:
        id:
          type: integer
        level:
          type: string
          enum: [info, warning, error, success]
        message:
          type: string
        source:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - level
        - message
        - createdAt

    CreateLogRequest:
      type: object
      properties:
        level:
          type: string
          enum: [info, warning, error, success]
        message:
          type: string
        source:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
          additionalProperties: true
      required:
        - level
        - message

    MemoryItem:
      type: object
      properties:
        id:
          type: integer
        key:
          type: string
        value:
          type: string
        category:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - key
        - value
        - createdAt
        - updatedAt

    CreateMemoryRequest:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
        category:
          type: string
          nullable: true
      required:
        - key
        - value

    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [active, planned, blocked, on_hold, done]
        priority:
          type: string
          enum: [critical, high, medium, low]
        owner:
          type: string
          nullable: true
        nextStep:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - status
        - priority
        - createdAt
        - updatedAt

    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [active, planned, blocked, on_hold, done]
          default: planned
        priority:
          type: string
          enum: [critical, high, medium, low]
          default: medium
        owner:
          type: string
          nullable: true
        nextStep:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date
          nullable: true
      required:
        - name

    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [active, planned, blocked, on_hold, done]
        priority:
          type: string
          enum: [critical, high, medium, low]
        owner:
          type: string
          nullable: true
        nextStep:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date
          nullable: true

    UpdateProjectParams:
      type: object
      properties:
        id:
          type: integer
      required:
        - id

    DeleteProjectParams:
      type: object
      properties:
        id:
          type: integer
      required:
        - id

    CalendarEvent:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        date:
          type: string
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - date
        - createdAt
        - updatedAt

    CreateEventRequest:
      type: object
      properties:
        title:
          type: string
        date:
          type: string
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
      required:
        - title
        - date

    UpdateEventRequest:
      type: object
      properties:
        title:
          type: string
        date:
          type: string
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        location:
          type: string
          nullable: true

    InboxItem:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        processed:
          type: boolean
        projectId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - content
        - processed
        - createdAt

    CreateInboxItemRequest:
      type: object
      properties:
        content:
          type: string
        projectId:
          type: integer
          nullable: true
      required:
        - content

    UpdateInboxItemRequest:
      type: object
      properties:
        processed:
          type: boolean
        projectId:
          type: integer
          nullable: true