# lib/api-zod/src/generated/types/calendarEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mission Control API for openclaw agent Ben
 * OpenAPI spec version: 0.1.0
 */

export interface CalendarEvent {
  id: number;
  title: string;
  date: string;
  startTime?: string | null;
  endTime?: string | null;
  description?: string | null;
  location?: string | null;
  createdAt: Date;
  updatedAt: Date;
}