# lib/api-zod/src/generated/types/createEventRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mission Control API for openclaw agent Ben
 * OpenAPI spec version: 0.1.0
 */

export interface CreateEventRequest {
  title: string;
  date: string;
  startTime?: string | null;
  endTime?: string | null;
  description?: string | null;
  location?: string | null;
}