# lib/api-zod/src/generated/types/createInboxItemRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mission Control API for openclaw agent Ben
 * OpenAPI spec version: 0.1.0
 */

export interface CreateInboxItemRequest {
  content: string;
  projectId?: number | null;
}