# lib/api-zod/src/generated/types/inboxItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mission Control API for openclaw agent Ben
 * OpenAPI spec version: 0.1.0
 */

export interface InboxItem {
  id: number;
  content: string;
  processed: boolean;
  projectId?: number | null;
  createdAt: Date;
}