# lib/api-zod/src/generated/types/updateInboxItemRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mission Control API for openclaw agent Ben
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateInboxItemRequest {
  processed?: boolean;
  projectId?: number | null;
}