# lib/api-zod/src/generated/types/updateTaskRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mission Control API for openclaw agent Ben
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTaskRequestPriority } from "./updateTaskRequestPriority";
import type { UpdateTaskRequestStatus } from "./updateTaskRequestStatus";

export interface UpdateTaskRequest {
  title?: string;
  description?: string | null;
  status?: UpdateTaskRequestStatus;
  priority?: UpdateTaskRequestPriority;
  dueDate?: string | null;
  projectId?: number | null;
}